+++ packages/akashic-cli-serve/src/client/operator/PlayOperator.ts
@@ -26,15 +26,17 @@ export class PlayOperator {
 
 	openNewClientInstance = (): void => {
 		// ignoreSession は Mac Chrome の不具合(？)対策でやむなくつけているフラグ。 (ref. ../store/storage.ts)
+		console.log(`OPENWINDOW: width=${window.innerWidth},height=${window.innerHeight},noopener`);
 		window.open(
-			`${window.location.pathname}?id=${this.store.contentLocator.contentId}&ignoreSession=1`,
+			`${window.location.pathname}?ignoreSession=1`,
 			"_blank",
 			`width=${window.innerWidth},height=${window.innerHeight},noopener`
 		);
 	}
 
 	sendRegisteredEvent = (eventName: string): void => {
-		const pevs = this.store.sandboxConfig.events[eventName];
+		const sandboxConfig = this.store.currentLocalInstance.content.sandboxConfig || {};
+		const pevs = sandboxConfig.events[eventName];
 		const amflow = this.store.currentPlay.amflow;
 		pevs.forEach((pev: any) => amflow.sendEvent(pev));
 	}