+++ packages/akashic-cli-serve/src/client/store/storage.ts
@@ -32,6 +32,10 @@ function choose<T>(a: T | null | undefined, b: T | null | undefined, c: T): T {
 	return (a != null) ? a : ((b != null) ? b : c);
 }
 
+function getQueryValue(queryValue: string | string[]): string {
+	return queryValue as string; // TODO クエリの値が配列の場合の対応を考える
+}
+
 export class Storage {
 	static SESSION_STORAGE_KEY: string = "aktb:config";
 
@@ -61,20 +65,20 @@ export class Storage {
 			}
 		}
 
-		const playerId =  choose(qp.playerId, s.playerId, ("" + Date.now())); // TODO 名前も含めサーバに問い合わせる？
+		const playerId =  choose(getQueryValue(qp.playerId), s.playerId, ("" + Date.now())); // TODO 名前も含めサーバに問い合わせる？
 		this.put({
 			playerId,
-			playerName: choose(qp.playerName, s.playerName, ("player" + playerId)),
-			showsDevtools: choose(asBool(qp.showsDevtools), s.showsDevtools, false),
-			devtoolsHeight: choose(asNumber(qp.devtoolsHeight), s.devtoolsHeight, 200),
-			activeDevtool: choose(qp.activeDevtool, s.activeDevtool, "Instances"),
-			showsEventList: choose(asBool(qp.showsEventList), s.showsEventList, true),
-			eventListWidth: choose(asNumber(qp.eventListWidth), s.eventListWidth, 150),
-			eventEditContent: choose(qp.eventEditContent, s.eventEditContent, ""),
-			selectedArgumentName: choose(qp.selectedArgumentName, s.selectedArgumentName, null),
-			instanceArgumentListWidth: choose(asNumber(qp.instanceArgumentListWidth), s.instanceArgumentListWidth, 150),
-			instanceArgumentEditContent: choose(qp.instanceArgumentEditContent, s.instanceArgumentEditContent, ""),
-			joinsAutomatically: choose(asBool(qp.joinsAutomatically), s.joinsAutomatically, false)
+			playerName: choose(getQueryValue(qp.playerName), s.playerName, ("player" + playerId)),
+			showsDevtools: choose(asBool(getQueryValue(qp.showsDevtools)), s.showsDevtools, false),
+			devtoolsHeight: choose(asNumber(getQueryValue(qp.devtoolsHeight)), s.devtoolsHeight, 200),
+			activeDevtool: choose(getQueryValue(qp.activeDevtool), s.activeDevtool, "Instances"),
+			showsEventList: choose(asBool(getQueryValue(qp.showsEventList)), s.showsEventList, true),
+			eventListWidth: choose(asNumber(getQueryValue(qp.eventListWidth)), s.eventListWidth, 150),
+			eventEditContent: choose(getQueryValue(qp.eventEditContent), s.eventEditContent, ""),
+			selectedArgumentName: choose(getQueryValue(qp.selectedArgumentName), s.selectedArgumentName, null),
+			instanceArgumentListWidth: choose(asNumber(getQueryValue(qp.instanceArgumentListWidth)), s.instanceArgumentListWidth, 150),
+			instanceArgumentEditContent: choose(getQueryValue(qp.instanceArgumentEditContent), s.instanceArgumentEditContent, ""),
+			joinsAutomatically: choose(asBool(getQueryValue(qp.joinsAutomatically)), s.joinsAutomatically, false)
 		});
 	}