+++ packages/akashic-cli-serve/src/client/view/App.tsx
@@ -25,10 +25,11 @@ export class App extends React.Component<AppProps, {}> {
 				<StartupScreenContainer
 					operator={operator}
 					startupScreenUiStore={store.startupScreenUiStore}
-					argumentsTable={store.argumentsTable}
+					argumentsTable={store.currentPlay.content.argumentsTable}
 				/>
 			</div>;
 		}
+		const sandboxConfig = store.currentLocalInstance.content.sandboxConfig || {};
 		return <div id="whole" className={styles["whole"]}>
 			<ToolBarContainer
 				play={store.currentPlay}
@@ -39,7 +40,7 @@ export class App extends React.Component<AppProps, {}> {
 			<div className={styles["main"] + " " + styles["centering"] } ref={this._onRef}>
 			{
 				store.toolBarUiStore.showsBgImage ?
-					<img src={this.props.store.sandboxConfig.backgroundImage} className={styles["bg-image"]}/> :
+					<img src={sandboxConfig.backgroundImage} className={styles["bg-image"]}/> :
 					null
 			}
 			</div>
@@ -50,7 +51,7 @@ export class App extends React.Component<AppProps, {}> {
 							play={store.currentPlay}
 							operator={operator}
 							devtoolUiStore={store.devtoolUiStore}
-							sandboxConfig={store.sandboxConfig}
+							sandboxConfig={sandboxConfig}
 						/>
 					</div> :
 					null