+++ packages/akashic-cli-serve/src/server/controller/ScriptAssetController.ts
@@ -11,10 +11,9 @@ export const createScriptAssetController = (baseDir: string): express.RequestHan
 	const watcher = chokidar.watch(gameJsonPath, { persistent: true });
 	watcher.on("change", () => {
 		try {
-			const gameJsonFileValue = JSON.parse(fs.readFileSync(gameJsonPath).toString());
-			gameJson = gameJsonFileValue;
+			gameJson = JSON.parse(fs.readFileSync(gameJsonPath).toString());
 		} catch (e) {
-			getSystemLogger().warn("detected game.json updated but dit not reflect because parsing failed.");
+			// do nothing
 		}
 	});
 
@@ -27,6 +26,7 @@ export const createScriptAssetController = (baseDir: string): express.RequestHan
 			return;
 		}
 		let id = Object.keys(gameJson.assets).find((id) => gameJson.assets[id].path === req.params.scriptName);
+
 		const content = fs.readFileSync(scriptPath);
 		const responseBody = `"use strict";
 			if (! ("gScriptContainer" in window)) {