+++ packages/akashic-cli-serve/src/server/domain/dynamicRequire.ts
@@ -1,0 +1,10 @@
+export function dynamicRequire<T>(path: string): T | null {
+	let ret: T | null = null;
+	try {
+		ret = require(path) as T;
+	} catch (e) {
+		if (e.code !== "MODULE_NOT_FOUND")
+			throw e;
+	}
+	return ret;
+}