+++ CHANGELOG.md
@@ -1,5 +1,33 @@
 # CHANGELOG
 
+## 1.7.2 (2019-07-02)
+
+#### Other Change
+* [#94](https://github.com/akashic-games/akashic-cli/pull/94) 何も変更を加えずにpublishするnpm-scriptを追加 ([@dera-](https://github.com/dera-))
+
+#### Update Dependencies
+* `akashic-cli-commons`, `akashic-cli-config`, `akashic-cli-export-html`, `akashic-cli-export-zip`, `akashic-cli-init`, `akashic-cli-install`, `akashic-cli-modify`, `akashic-cli-scan`, `akashic-cli-serve`, `akashic-cli-stat`, `akashic-cli-uninstall`, `akashic-cli-update`, `akashic-cli`
+  * [#106](https://github.com/akashic-games/akashic-cli/pull/106) Update all dependencies (patch) ([@renovate[bot]](https://github.com/apps/renovate))
+
+#### Republish
+* [#111](https://github.com/akashic-games/akashic-cli/pull/111) Republish ([@xnv](https://github.com/xnv))
+
+#### Committers: 2
+- [@dera-](https://github.com/dera-)
+- xnv ([@xnv](https://github.com/xnv))
+
+## 1.7.1 (2019-06-28)
+
+#### Enhancement
+* `akashic-cli-serve`
+  * [#110](https://github.com/akashic-games/akashic-cli/pull/110) 【akashic-cli-serve】内部コンポーネントの更新(engineFiles@2.1.19、engineFiles@1.1.9) ([@xnv](https://github.com/xnv))
+* `akashic-cli-export-zip`
+  * [#109](https://github.com/akashic-games/akashic-cli/pull/109) export-htmlでatsumaruオプション実行時もgame.jsonにexport-zip実行時の情報が出力されるようにする ([@dera-](https://github.com/dera-))
+
+#### Committers: 2
+- [@dera-](https://github.com/dera-)
+- xnv ([@xnv](https://github.com/xnv))
+
 ## 1.7.0 (2019-06-24)
 
 #### Breaking Change